+++ CHANGELOG.rst
@@ -13,12 +13,15 @@ Version 0.2.0
       zenodo for DOI generation.
     - Use python :code:`logging` module and remove inheritance from
       :code:`Base` class.
-    - Added :code:`tolerance` as class instance variable of the
-      :code:`GlacierFowModel` class.
     - Use internal method :code:`self._iterate` to simluate years in order to
       reduce duplicate code.
     - Getter and setter methods for the :code:`plot` instance variable, which
       initializes or destroys the :code:`matplotlib.pyplot.figure`.
+    - Functionality to export the glacier layers and statistics of the model as
+      :code:`.csv` and :code:`.tif` using the :code:`self.export()` method.
+    - Reproject example DEM :code:`aletsch.tif` from Swiss CH1903 / LV03
+      (EPSG:21781) to Swiss CH1903+ / LV95 (EPSG:2056).
+    - Add flow and model parameters as class attributes.
 - Bugfixes:
     - Fix failing CI: Update package dependencies, set GitHub actions to python
       3.10, set GDAL version to 3.4.1 and remove shebang from tests.
@@ -27,6 +30,11 @@ Version 0.2.0
       will now perform a clean reset of the model.
     - A model destructor ensures closing the model figure, when the model is
       deleted or garbage collected.
+    - Random nudging had no effect, now the layer is considered if class
+      variable :code:`MODEL_RANDOM_NUDGING` is set to :code:`True` (default).
+    - Clarify the velocity variable :code:`ud` as surface ice deformation
+      velocity (at medium height), and point out that basal sliding and soft
+      bed deformation are ignored.
 
 Version 0.1.2
 -------------